+++ web/src/engine/websites/HiveScans.ts
@@ -1,11 +1,11 @@
 import { Tags } from '../Tags';
-import icon from './MangaPro.webp';
+import icon from './HiveScans.webp';
 import { Iken } from './templates/Iken';
 
 export default class extends Iken {
 
     public constructor() {
-        super('mangapro', 'Manga Pro', 'https://promanga.pro', Tags.Media.Manga, Tags.Media.Manhwa, Tags.Media.Manhua, Tags.Language.Arabic, Tags.Source.Scanlator);
+        super('hivescans', 'Hive Scans', 'https://hivetoon.com', Tags.Media.Manhwa, Tags.Media.Manhua, Tags.Language.English, Tags.Source.Scanlator);
     }
 
     public override get Icon() {
