+++ web/src/engine/websites/_index.ts
@@ -198,6 +198,7 @@ export { default as Hinapyon } from './Hinapyon';
 export { default as HiperCool } from './HiperCool';
 export { default as Hiperdex } from './Hiperdex';
 export { default as Hitomi } from './Hitomi';
+export { default as HiveScans } from './HiveScans';
 export { default as HniScanTrad } from './HniScanTrad';
 export { default as HoiFansub } from './HoiFansub';
 export { default as HorrorFC } from './HorrorFC';
@@ -690,7 +691,6 @@ export { default as VerMangasPorno } from './VerMangasPorno';
 export { default as VerManhwa } from './VerManhwa';
 export { default as ViewComics } from './ViewComics';
 export { default as VNSharing } from './VNSharing';
-export { default as VoidScans } from './VoidScans';
 export { default as VortexScans } from './VortexScans';
 export { default as VyManga } from './VyManga';
 export { default as WebAce } from './WebAce';
