+++ web/src/engine/websites/templates/Iken_e2e.ts
@@ -1,0 +1,3 @@
+import '../HiveScans_e2e';
+import '../MangaPro_e2e';
+import '../VortexScans_e2e';
